+++ src/service/mod.rs
@@ -1,3 +1,2 @@
-pub mod client;
 pub mod resource;
 pub mod table;
